role checking in bll started

+++ src/main/java/bll/manager/UserManager.java
@@ -1,15 +1,20 @@
 package bll.manager;
 
 import be.User;
+import be.enums.UserRole;
 import bll.IManager;
 import dal.DAOFactory;
 import dal.dao.UserDAO;
+import gui.model.UserModel;
+import utils.permissions.Checker;
+import utils.permissions.RequiresPermission;
 
 import java.util.Map;
 import java.util.UUID;
 
 public class UserManager implements IManager<User> {
     private UserDAO dao;
+    private Checker checker;
 
     public UserManager() {
         dao = (UserDAO) DAOFactory.createDAO(DAOFactory.DAOType.USER);
@@ -26,8 +31,15 @@ public class UserManager implements IManager<User> {
     }
 
     @Override
+    @RequiresPermission(UserRole.ADMINISTRATOR)
     public String delete(UUID id) {
-        return dao.delete(id);
+
+        if (id.equals(UserModel.getInstance().getLoggedInUser().getUserID())
+                || checker.hasAccess(this.getClass())) {
+            return dao.delete(id);
+        }
+        else {
+            throw new SecurityException("You do not have permission to delete this user.");}
     }
 
     @Override
